requirements.txt
asgiref==3.3.4
Babel==2.9.1
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cssselect2==0.4.1
defusedxml==0.7.1
dj-static==0.0.6
Django==3.2.4
django-anymail==8.4
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-pdfkit==0.3.1
django-phonenumber-field==5.2.0
django-renderpdf==3.0.1
django-rest-passwordreset==1.2.0
django-tinymce==3.3.0
djangorestframework==3.12.4
gunicorn==20.1.0
html5lib==1.1
idna==2.10
mysqlclient==2.0.3
pdfkit==0.6.0
phonenumberslite==8.12.25
Pillow==8.2.0
psycopg2==2.9.1
pycparser==2.20
pyphen==0.11.0
pytz==2021.1
reportlab==3.5.67
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
tinycss2==1.1.0
urllib3==1.26.5
WeasyPrint==52.5
webencodings==0.5.1
XlsxWriter==1.4.3
